fix(stage): correct api url path
