+++ src/commands/index.ts
@@ -27,9 +27,9 @@ import { resolveJudgeCommandButtonEvents, useJudge } from './judge'
 import { useForumClose } from './forum/forum_close'
 import { useForumCreate } from './forum/forum_create'
 import { useStageATA } from './stage/stage_ata'
-import { useStageEnter } from './stage/stage_enter'
-import { useStageFinish } from './stage/stage_finish'
 import { useStageStart } from './stage/stage_start'
+import { useStageFinish } from './stage/stage_finish'
+import { useStageEnter } from './stage/stage_enter'
 
 const registerHooks = (client: He4rtClient, commands: Command[]) => {
   commands.forEach(([data, cb]) => {
@@ -64,10 +64,10 @@ export const registerCommands = async ({ client, rest }: Context) => {
     useJudge(),
     useForumClose(),
     useForumCreate(),
-    // useStageATA(),
-    // useStageEnter(),
-    // useStageFinish(),
-    // useStageStart(),
+    useStageATA(),
+    useStageEnter(),
+    useStageFinish(),
+    useStageStart(),
     // useReputation()
   ])
