+++ src/commands/stage/stage_enter.ts
@@ -1,30 +1,26 @@
-import { SlashCommandBuilder } from 'discord.js'
-import { Command, MeetingPOST } from '@/types'
+import { PermissionFlagsBits, SlashCommandBuilder } from 'discord.js'
+import { Command, MeetingAttendPost } from '@/types'
 import { STAGE } from '@/defines/commands.json'
 import { MEETING_OPTION } from '-/commands/stage_enter.json'
-import { getOption, reply } from '@/utils'
+import { reply } from '@/utils'
 
 export const useStageEnter = (): Command => {
   const data = new SlashCommandBuilder()
     .setName(STAGE.TITLE)
     .setDescription(STAGE.DESCRIPTION)
     .setDMPermission(false)
-    .addStringOption((option) => option.setName('reuniao').setDescription(MEETING_OPTION).setRequired(true))
+    .addIntegerOption((option) => option.setName('reuniao').setDescription(MEETING_OPTION).setRequired(true))
+    .setDefaultMemberPermissions(PermissionFlagsBits.Administrator)
 
   return [
     data,
     async (interaction, client) => {
-      const meeting = getOption(interaction, 'reuniao')
-
-      const meeting_type_id = meeting.value as number
-
-      client.api.he4rt.meeting
-        .post<MeetingPOST>({
-          meeting_type_id,
+      client.api.he4rt.events.meeting.attend
+        .post<MeetingAttendPost>({
           discord_id: interaction.user.id,
         })
-        .then(async () => {
-          await reply(interaction).success()
+        .then(async ({ meeting_id }) => {
+          await interaction.reply(`A reunião foi criada! O id da reunião é **${meeting_id}**!`)
         })
         .catch(async () => {
           await reply(interaction).error()
