+++ src/commands/stage/stage_start.ts
@@ -1,26 +1,30 @@
-import { PermissionFlagsBits, SlashCommandBuilder } from 'discord.js'
-import { Command, MeetingAttendPost } from '@/types'
+import { SlashCommandBuilder } from 'discord.js'
+import { Command, MeetingPOST } from '@/types'
 import { STAGE_START } from '@/defines/commands.json'
 import { MEETING_OPTION } from '-/commands/stage_enter.json'
-import { reply } from '@/utils'
+import { getOption, reply } from '@/utils'
 
 export const useStageStart = (): Command => {
   const data = new SlashCommandBuilder()
     .setName(STAGE_START.TITLE)
     .setDescription(STAGE_START.DESCRIPTION)
     .setDMPermission(false)
-    .addStringOption((option) => option.setName('reuniao').setDescription(MEETING_OPTION).setRequired(true))
-    .setDefaultMemberPermissions(PermissionFlagsBits.Administrator)
+    .addIntegerOption((option) => option.setName('reuniao').setDescription(MEETING_OPTION).setRequired(true))
 
   return [
     data,
     async (interaction, client) => {
-      client.api.he4rt.meeting.attend
-        .post<MeetingAttendPost>({
+      const meeting = getOption(interaction, 'reuniao')
+
+      const meeting_type_id = meeting.value as number
+
+      client.api.he4rt.events.meeting
+        .post<MeetingPOST>({
+          meeting_type_id,
           discord_id: interaction.user.id,
         })
-        .then(async ({ meeting_id }) => {
-          await interaction.reply(`A reunião foi criada! O id da reunião é **${meeting_id}**!`)
+        .then(async () => {
+          await reply(interaction).success()
         })
         .catch(async () => {
           await reply(interaction).error()
